+++ cmd/hidden_lake/service/internal/handler/node_key.go
@@ -12,7 +12,7 @@ import (
 	http_logger "github.com/number571/go-peer/internal/logger/http"
 )
 
-func HandleNodeKeyAPI(pWrapper config.IWrapper, pLogger logger.ILogger, pNode anonymity.INode) http.HandlerFunc {
+func HandleNetworkPubKeyAPI(pWrapper config.IWrapper, pLogger logger.ILogger, pNode anonymity.INode) http.HandlerFunc {
 	return func(pW http.ResponseWriter, pR *http.Request) {
 		logBuilder := http_logger.NewLogBuilder(pkg_settings.CServiceName, pR)