+++ cmd/hidden_lake/traffic/pkg/client/types.go
@@ -1,11 +1,13 @@
 package client
 
 import (
+	"github.com/number571/go-peer/cmd/hidden_lake/traffic/pkg/config"
 	net_message "github.com/number571/go-peer/pkg/network/message"
 )
 
 type IClient interface {
 	GetIndex() (string, error)
+	GetSettings() (config.IConfigSettings, error)
 
 	GetPointer() (uint64, error)
 	GetHash(uint64) (string, error)
@@ -20,6 +22,7 @@ type IBuilder interface {
 
 type IRequester interface {
 	GetIndex() (string, error)
+	GetSettings() (config.IConfigSettings, error)
 
 	GetPointer() (uint64, error)
 	GetHash(uint64) (string, error)